This commit adds support for when pipelines may not be speciallized
I was trying to use this crate together with [this crate](https://github.com/mattatz/bevy_points) and found this
line panicking. There may be times where a user adds objects that cannot
produce speciallized pipelines (for instance in this case, the crate
does not populate normals). It seems to me there is no need to panic in
these cases and just move on.